Improve text placement for line placement.
Instead of taking the line height (which includes line spacing) take the
height of character "X" which results in a reasonable height in almost all fonts.

Also change all integer values for text placement to double to improve accuracy.

+++ include/mapnik/font_engine_freetype.hpp
@@ -147,10 +147,10 @@ class MAPNIK_DECL font_face_set : private boost::noncopyable
 public:
     class dimension_t {
     public:
-        dimension_t(unsigned width_, int ymax_, int ymin_, double linespacing_) :  width(width_), height(ymax_-ymin_), linespacing(linespacing_), ymin(ymin_) {}
-        unsigned width, height;
+        dimension_t(double width_, double ymax_, double ymin_, double linespacing_) :  width(width_), height(ymax_-ymin_), linespacing(linespacing_), ymin(ymin_) {}
+        double width, height;
         double linespacing;
-        int ymin;
+        double ymin;
     };
 
     font_face_set(void)
